refactor: kebab-case filenames
`unicorn/filename-case` is on at kebabCase in the shared preset. Applied
with `magic-kebab --write`: 5 renames, 6 import specifiers rewritten, 0
skipped, 0 conflicts, 0 needing review.

`example/App.tsx` stays as it is — the react-native preset exempts it and
the codemod skips it, because Expo's `AppEntry.js` imports `../../App` from
inside node_modules where nothing can rewrite it.
