+++ src/@types/magic-toast-props.ts
