+++ src/assets/success-icon.tsx
