+++ src/components/alert-toast.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { StatusBar } from 'react-native';
-import { AlertIcon } from '../assets/AlertIcon';
+import { AlertIcon } from '../assets/alert-icon';
 import { Toast } from './Toast';
 
-import type { MagicToastProps } from '../@types/MagicToastProps';
+import type { MagicToastProps } from '../@types/magic-toast-props';
 
 export const AlertToast: React.FC<MagicToastProps> = ({
   message,
